Error Page implemented + FIXES
Implemented a error page that shows custom error titles, codes and messages.

Fixed user deletion endpoint: deleted all upvoted and downvoted posts.

+++ backend/src/main/java/com/example/backend/controller/APIUserController.java
@@ -494,6 +494,11 @@ public class APIUserController {
             return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
 
+        // Is the user the admin of any community?
+        if (userService.getNumberOfAdminCommunities(username) > 0) {
+            return new ResponseEntity<>("User is the admin of one or more communities", HttpStatus.UNAUTHORIZED);
+        }
+
         // Remove references to the user in other entities
         userService.removeUserReferences(user);
